consolidate query param types
