+++ clients/admin-ui/src/features/consent-settings/tcf/tcf-config.slice.ts
@@ -10,7 +10,7 @@ import {
   TCFPublisherRestrictionRequest,
   TCFPublisherRestrictionUpdateRequest,
 } from "~/types/api";
-import { PaginationQueryParams } from "~/types/common/PaginationQueryParams";
+import { PaginationQueryParams } from "~/types/query-params";
 
 export interface State {
   page: number;
