+++ clients/admin-ui/src/features/manual-tasks/manual-tasks.slice.ts
@@ -11,7 +11,7 @@ import {
   ManualFieldSearchResponse,
   ManualFieldStatus,
 } from "~/types/api";
-import { PaginationQueryParams } from "~/types/common/PaginationQueryParams";
+import { PaginationQueryParams } from "~/types/query-params";
 
 import { PAGE_SIZES } from "../common/table/v2";
