--- clients/admin-ui/src/types/common/DateRange.ts
@@ -1,6 +1,0 @@
-import { Dayjs } from "dayjs";
-
-export type DateRange = {
-  startDate?: Dayjs | null;
-  endDate?: Dayjs | null;
-};
