--- clients/admin-ui/src/types/common/PaginationQueryParams.ts
@@ -1,12 +1,0 @@
-export interface PaginationQueryParams {
-  page: number;
-  size: number;
-}
-
-export interface PaginatedResponse<T> {
-  items: T[];
-  page: number;
-  size: number;
-  total: number;
-  pages: number;
-}
