--- clients/admin-ui/src/types/common/SearchQueryParams.ts
@@ -1,3 +1,0 @@
-export interface SearchQueryParams {
-  search?: string;
-}
