--- clients/admin-ui/src/types/common/SortQueryParams.ts
@@ -1,4 +1,0 @@
-export interface SortQueryParams<T = string> {
-  sort_by?: T | T[];
-  sort_asc?: boolean;
-}
